Set values only when necessary.
This makes logs on client machines more usable. Until now, every time
the catalog was run, gconftool-2 was executed to set the gconf values
and this action was logged by puppet. Now, gconftool-2 is executed to
set the value only when current value is different, and if it is the
same, nothing is logged.